src/requirements.txt
asgiref==3.6.0
contourpy==1.0.6
cycler==0.11.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
